Revert the BSN screen-binding architecture; keep the principled wins
The bind-time RetailTree lookup was not architectural debt; replacing it with
generated XxxBindings components added a metadata layer in the generator, flat
row-field reconstruction in binders, and shape dispatch in lifecycle without
removing a real problem. Drop _VIEW_BINDINGS, the generated binding components,
the per-shape industry binders, and the Warehouse scalar reconstruction, and
restore the compact FourCC binding tables. Keep: integer evidence stays i32
end-to-end, the isolated BSN named-ref resolution test, and an explicit
Armory picture-pair API. The VecTemplate probe shows BSN refs only resolve as
scalar Entity fields, so repeated fixed controls keep their FourCC tables.

+++ rust/crates/imperialism-app/src/ui/city/dialogs/shipyard.rs
@@ -64,7 +64,7 @@ pub(in crate::ui::city) fn configure_shipyard_dialog(
         });
         if available {
             let (idle, active) =
-                shipyard_queue_pictures(assets, &queue_icons, slot, ship_type, overlay_left as i32);
+                shipyard_queue_pictures(assets, &queue_icons, slot, ship_type, overlay_left);
             commands.entity(button).insert((
                 RetailPictureSwap {
                     idle: idle.clone(),
@@ -243,7 +243,7 @@ pub(in crate::ui::city) fn render_shipyard_dialog(
             capabilities.resource_weight,
         ];
         for (index, &(left, baseline)) in generated::SHIPYARD_STAT_ORIGINS.iter().enumerate() {
-            let origin = IVec2::new(left as i32, baseline as i32);
+            let origin = IVec2::new(left, baseline);
             let label = city_string(&assets, 0x2736, 0x10 + index as i16);
             text.draw(&mut picture, origin, &label, 0xd2);
             text.draw(
